[Add] : implement Controller
